Cleared option/__init__.py  moved to options.pw
