+++ src/auto_flutter_dig/model/argument/arguments.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Dict, Iterable, Optional, Union
 
-from ...model.argument.option import LongOption, Option, OptionAll, PositionalOption, ShortOption
+from ...model.argument.options import LongOption, Option, OptionAll, PositionalOption, ShortOption
 
 Value = Optional[str]
 Argument = str
