--- src/auto_flutter_dig/model/argument/option/__init__.py
@@ -1,22 +1,0 @@
-from ....model.argument.option.all import OptionAll
-from ....model.argument.option.long import LongOption, LongOptionWithValue
-from ....model.argument.option.long_positional import LongPositionalOption
-from ....model.argument.option.long_short import LongShortOption, LongShortOptionWithValue
-from ....model.argument.option.option import Option
-from ....model.argument.option.positional import PositionalOption
-from ....model.argument.option.short import ShortOption, ShortOptionWithValue
-from ....model.argument.option.valued import OptionWithValue
-
-__all__ = [
-    "OptionAll",
-    "LongOption",
-    "LongOptionWithValue",
-    "LongPositionalOption",
-    "LongShortOption",
-    "LongShortOptionWithValue",
-    "Option",
-    "PositionalOption",
-    "ShortOption",
-    "ShortOptionWithValue",
-    "OptionWithValue",
-]
