+++ src/auto_flutter_dig/model/argument/option/all.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 
-from ....model.argument.option import Option
+from ....model.argument.option.option import Option
 
 __all__ = ["OptionAll"]
