+++ src/auto_flutter_dig/module/aflutter/task/config/project.py
@@ -1,6 +1,6 @@
 from typing import Callable
 
-from .....model.argument.option import Option
+from .....model.argument.options import Option
 from .....model.task.task import *
 from .....model.task.identity import TaskIdentity
