+++ src/auto_flutter_dig/module/aflutter/task/setup/check.py
@@ -1,5 +1,5 @@
 from .....core.utils import _Ensure, _Iterable
-from .....model.argument.option import LongOption, LongShortOption, Option
+from .....model.argument.options import LongOption, LongShortOption, Option
 from .....model.task.task import *
 from .....module.aflutter.identity import AflutterTaskIdentity, TaskIdentity
