+++ src/auto_flutter_dig/module/firebase/task/setup/setup.py
@@ -2,7 +2,7 @@ from .....core.config import Config
 from .....core.os.executable_resolver import ExecutableResolver
 from .....core.os.path_converter import PathConverter
 from .....core.string import SB
-from .....model.argument.option import LongOption, LongPositionalOption
+from .....model.argument.options import LongOption, LongPositionalOption
 from .....model.task.task import *
 from .....module.aflutter.task.setup.save import AflutterSetupSaveTask
 from .....module.firebase.identity import FirebaseTaskIdentity
