+++ src/auto_flutter_dig/module/firebase/task/upload.py
@@ -4,7 +4,7 @@ from ....core.config import Config
 from ....core.os.path_converter import PathConverter
 from ....core.utils import _Dict, _If
 from ....core.utils.task.process.process import *
-from ....model.argument.option import LongOptionWithValue
+from ....model.argument.options import LongOptionWithValue
 from ....model.error import E
 from ....module.firebase.identity import FirebaseTaskIdentity
 from ....module.firebase.model._const import FIREBASE_CONFIG_KEY_PATH, FIREBASE_DISABLE_INTERACTIVE_MODE, FIREBASE_ENV
